1. flutter项目视图排除windows、macos、linux平台目录 2. 生成Images引用资源时，删除不存在的图片引用时，除了使用虚拟文件系统，还增加了真实文件系统的判断

+++ src/main/kotlin/com/wanggaowan/tools/actions/image/GeneratorImageRefAction.kt
@@ -31,6 +31,7 @@ import org.jetbrains.yaml.psi.YAMLDocument
 import org.jetbrains.yaml.psi.YAMLMapping
 import org.jetbrains.yaml.psi.YAMLScalar
 import org.jetbrains.yaml.psi.YAMLSequence
+import java.io.File
 
 /**
  * 生成图片资源引用，并在pubspec.yaml中生成图片位置声明
@@ -244,7 +245,10 @@ object GeneratorImageRefUtils {
                 }
 
                 if (manager.findFileByUrl("file://${moduleRootFile.path}/$value") == null) {
-                    it.delete()
+                    val file= File("${moduleRootFile.path}/$value")
+                    if (!file.exists()) {
+                        it.delete()
+                    }
                 }
             }
         }
